Phase 6 (#93): retire dbus-java — the owned connection is now the JVM backend
The self-owned wire connection (junixsocket + our own marshaller + read/dispatch
loop) now does everything dbus-java did, proven against real external peers across
phases 1-5 (client calls, properties, signal reception + emission, serving incoming
calls, fd-passing). This finale makes it THE JVM backend and removes dbus-java.

Removed:
- the `-Dsdbus.jvm.wire` / `SDBUS_JVM_WIRE` toggle and `isJvmWireBackendEnabled`;
  JvmDbusBackendProvider.backend is now unconditionally WireDbusBackend.
- the dbus-java dependency from jvmMain (dbus-java-core, dbus-java-transport-
  junixsocket) and the version-catalog entries.
- PureJavaDbusBackend.kt (the dbus-java connection/proxy/object path + all
  dbus-java conversion/credential helpers + the in-process LocalJvmSignalBus) and
  LocalJvmMatchBus.kt (the dbus-java/stub same-process signal mechanism).
- the in-process StubJvmDbusBackend.
- dbus-java-era tests: StubJvmDbusBackendTest, PureJavaDbusBackendTest, and
  DBusMarshallerDifferentialTest (the dbus-java marshalling oracle, now redundant
  with the spec tests + real-peer round-trips).

Kept / migrated:
- junixsocket is now a direct jvmMain dependency (was transitive via dbus-java); the
  published -jvm POM keeps junixsocket-core and drops dbus-java.
- the wire stack untouched: DBusMarshaller, WireMessage/Codec, DBusWireConnection,
  WireDbusBackend, WireServe, the value-model bridge, JvmStaticDispatch.
- the shared object machinery (formerly PureJavaDbusObject) extracted to
  WireDbusObject.kt, wire-only, alongside the shared registries
  (LocalManagedObjectsRegistry, LocalObjectManagerRegistry, LocalJvmServiceRegistry).
- backend-boundary mock tests (JvmConnectionMockkTest, JvmFactoryMockkTest,
  JvmProxyObjectMockkTest) keep passing — they mock the JvmDbus* interfaces.
- wire-gated cross_test cases (WireServeExternalTest, the fd cross-tests) now run
  unconditionally; the dbusmock peer-parity flags (#71/#72) reflect the wire backend.

CI: removed the redundant `wire-client-parity-x64` job (the wire backend is now the
only backend the full-tests job exercises) and added the `timeout 1800` + `--no-daemon`
hard guard to the full-tests gradle run; python3-dbusmock stays installed.

Verified (JDK 21): apiDump is a no-op (public API unchanged); ktlintCheck + apiCheck
pass; full JVM suite green on the sole backend (297 :jvmTest + 60 :cross_test:jvmTest,
0 skipped, dbusmock suites engaged with real timings); compileKotlinLinuxX64 +
:cross_test:compileTestKotlinLinuxX64 compile.

#101 (bound the serve worker pool) left open: the current newCachedThreadPool grows
unbounded and so does not starve nested same-connection calls; a sized pool risks
deadlocking them, so it is deferred rather than risk destabilizing the green backend.

dbus-java remains only as an INDEPENDENT third-party D-Bus peer in cross_test /
stress_test interop tests (declared by literal coordinate, kept out of the version
catalog and the published artifacts).

Fixes #93.
