--- src/jvmMain/kotlin/com/monkopedia/sdbus/internal/jvmdbus/PureJavaDbusBackend.kt
@@ -1,1947 +1,0 @@
-package com.monkopedia.sdbus.internal.jvmdbus
-
-import com.monkopedia.sdbus.ActionResource
-import com.monkopedia.sdbus.BusName
-import com.monkopedia.sdbus.CloseableResource
-import com.monkopedia.sdbus.Connection
-import com.monkopedia.sdbus.InterfaceName
-import com.monkopedia.sdbus.JvmConnection
-import com.monkopedia.sdbus.MemberName
-import com.monkopedia.sdbus.Message
-import com.monkopedia.sdbus.MessageHandler
-import com.monkopedia.sdbus.MethodCall
-import com.monkopedia.sdbus.MethodName
-import com.monkopedia.sdbus.MethodReply
-import com.monkopedia.sdbus.MethodVTableItem
-import com.monkopedia.sdbus.ObjectPath
-import com.monkopedia.sdbus.PendingAsyncCall
-import com.monkopedia.sdbus.PlainMessage
-import com.monkopedia.sdbus.PropertyGetReply
-import com.monkopedia.sdbus.PropertyName
-import com.monkopedia.sdbus.PropertySetCall
-import com.monkopedia.sdbus.PropertyVTableItem
-import com.monkopedia.sdbus.Resource
-import com.monkopedia.sdbus.ServiceName
-import com.monkopedia.sdbus.Signal
-import com.monkopedia.sdbus.SignalHandler
-import com.monkopedia.sdbus.SignalName
-import com.monkopedia.sdbus.Signature
-import com.monkopedia.sdbus.UnixFd
-import com.monkopedia.sdbus.VTableItem
-import com.monkopedia.sdbus.Variant
-import com.monkopedia.sdbus.createError
-import com.monkopedia.sdbus.methodReplyFrom
-import com.monkopedia.sdbus.signalFromMetadata
-import com.sun.security.auth.module.UnixSystem
-import java.nio.file.Files
-import java.nio.file.Paths
-import java.util.concurrent.atomic.AtomicBoolean
-import java.util.concurrent.atomic.AtomicLong
-import java.util.concurrent.atomic.AtomicReference
-import kotlin.concurrent.thread
-import kotlin.time.Duration
-import org.freedesktop.dbus.DBusMatchRule
-import org.freedesktop.dbus.FileDescriptor
-import org.freedesktop.dbus.connections.AbstractConnection
-import org.freedesktop.dbus.connections.impl.DBusConnection
-import org.freedesktop.dbus.connections.impl.DBusConnectionBuilder
-import org.freedesktop.dbus.connections.impl.DirectConnectionBuilder
-import org.freedesktop.dbus.interfaces.DBus
-import org.freedesktop.dbus.interfaces.DBusSigHandler
-import org.freedesktop.dbus.messages.DBusSignal
-import org.freedesktop.dbus.types.UInt16
-import org.freedesktop.dbus.types.UInt32
-import org.freedesktop.dbus.types.UInt64
-
-private const val OBJECT_MANAGER_INTERFACE_NAME = "org.freedesktop.DBus.ObjectManager"
-private const val OBJECT_MANAGER_GET_MANAGED_OBJECTS = "GetManagedObjects"
-
-internal class PureJavaDbusBackend : JvmDbusBackend {
-    override fun createConnection(
-        busType: JvmBusType,
-        endpoint: String?,
-        name: ServiceName?,
-        fd: Int?
-    ): JvmDbusConnection {
-        if (busType == JvmBusType.DIRECT_FD) {
-            throw createError(
-                -1,
-                "JVM backend does not support createDirectBusConnection(fd)"
-            )
-        }
-        if (busType == JvmBusType.SERVER_FD) {
-            throw createError(
-                -1,
-                "JVM backend does not support createServerBusConnection(fd)"
-            )
-        }
-        // Native parity (issue #81): an unreachable bus must surface as a thrown Error, like
-        // the native backend's openBus ("Failed to open bus", errno from sd_bus_open_*) --
-        // never a silent fallback to an in-process stub that fakes success. dbus-java does
-        // not expose an errno, so the exact error *name* is not pinned across backends, but
-        // the Error type and the throw are.
-        val connection = runCatching {
-            tryCreateConnection(busType, endpoint, fd)
-        }.getOrElse {
-            throw createError(
-                -1,
-                "Failed to open bus${endpoint?.let { e -> " '$e'" }.orEmpty()}: ${it.message}"
-            )
-        } ?: throw createError(
-            -1,
-            "Failed to open bus: unsupported connection request for $busType"
-        )
-        return PureJavaDbusConnection(connection).also { conn ->
-            if (name != null) {
-                conn.requestName(name)
-            }
-        }
-    }
-
-    override fun createProxy(
-        connection: Connection,
-        destination: ServiceName,
-        objectPath: ObjectPath,
-        runEventLoopThread: Boolean
-    ): JvmDbusProxy {
-        if (runEventLoopThread) {
-            connection.startEventLoop()
-        }
-        val connectionBackend = (connection as? JvmConnection)?.backend
-        val javaConnection = connectionBackend as? PureJavaDbusConnection
-        return PureJavaDbusProxy(
-            javaConnection?.javaConnection,
-            destination,
-            objectPath,
-            runCatching { connection.uniqueName.value }.getOrNull(),
-            isConnectionReleased = { javaConnection?.isReleased() == true },
-            defaultTimeoutMicros = {
-                connectionBackend?.getMethodCallTimeout()
-                    ?.inWholeMicroseconds
-                    ?.coerceAtLeast(0L)
-                    ?.toULong()
-                    ?: 0uL
-            }
-        )
-    }
-
-    override fun createObject(connection: Connection, objectPath: ObjectPath): JvmDbusObject =
-        PureJavaDbusObject(
-            objectPath,
-            (connection as? JvmConnection)?.backend
-                .let { it as? PureJavaDbusConnection }
-                ?.javaConnection,
-            runCatching { connection.uniqueName.value }.getOrNull()
-        )
-
-    private fun tryCreateConnection(
-        busType: JvmBusType,
-        endpoint: String?,
-        fd: Int?
-    ): AbstractConnection? {
-        if (fd != null) return null
-        return when (busType) {
-            JvmBusType.DEFAULT,
-            JvmBusType.SESSION -> DBusConnectionBuilder.forSessionBus().withShared(false).build()
-            JvmBusType.SYSTEM -> DBusConnectionBuilder.forSystemBus().withShared(false).build()
-            JvmBusType.SESSION_ADDRESS ->
-                endpoint?.let(DBusConnectionBuilder::forAddress)?.withShared(false)?.build()
-            JvmBusType.DIRECT_ADDRESS -> endpoint?.let(DirectConnectionBuilder::forAddress)?.build()
-            JvmBusType.DIRECT_FD,
-            JvmBusType.SERVER_FD -> null
-        }
-    }
-}
-
-private data class MatchFilter(
-    val sender: String? = null,
-    val path: String? = null,
-    val interfaceName: String? = null,
-    val member: String? = null
-)
-
-private object LocalJvmSignalBus {
-    data class Key(
-        val destination: String,
-        val path: String,
-        val interfaceName: String,
-        val member: String
-    )
-
-    private val nextId = AtomicLong(1)
-    private val handlers = mutableMapOf<Key, MutableMap<Long, SignalHandler>>()
-
-    fun register(
-        destination: String,
-        path: String,
-        interfaceName: String,
-        member: String,
-        signalHandler: SignalHandler
-    ): Resource {
-        val id = nextId.getAndIncrement()
-        val key = Key(destination, path, interfaceName, member)
-        synchronized(this) {
-            val perKey = handlers.getOrPut(key) { mutableMapOf() }
-            perKey[id] = signalHandler
-        }
-        return ActionResource {
-            synchronized(this) {
-                val perKey = handlers[key] ?: return@synchronized
-                perKey.remove(id)
-                if (perKey.isEmpty()) {
-                    handlers.remove(key)
-                }
-            }
-        }
-    }
-
-    fun emit(
-        sender: String?,
-        path: String,
-        interfaceName: String,
-        member: String,
-        payload: List<Any?>
-    ) {
-        LocalJvmMatchBus.emit(sender, path, interfaceName, member, payload)
-        val message = signalFromMetadata(
-            Message.Metadata(
-                interfaceName = interfaceName,
-                memberName = member,
-                sender = sender,
-                path = path,
-                valid = true,
-                empty = payload.isEmpty()
-            )
-        ).also {
-            it.payload.addAll(payload)
-        }
-        val callbacks = synchronized(this) {
-            handlers.filterKeys { key ->
-                key.path == path &&
-                    key.interfaceName == interfaceName &&
-                    key.member == member &&
-                    (key.destination.isEmpty() || key.destination == sender)
-            }.values.flatMap { it.values }.toList()
-        }
-        callbacks.forEach { callback ->
-            JvmCurrentMessageContext.withMessage(message) {
-                callback(message)
-            }
-        }
-    }
-}
-
-private object LocalManagedObjectsRegistry {
-    private data class ManagedInterfaceEntry(
-        val providers: MutableMap<Long, () -> Map<PropertyName, Variant>> = mutableMapOf()
-    )
-
-    private data class ManagedObjectKey(
-        val destination: String,
-        val objectPath: String,
-        val interfaceName: String
-    )
-
-    private val nextProviderId = AtomicLong(1)
-    private val interfaces = mutableMapOf<ManagedObjectKey, ManagedInterfaceEntry>()
-
-    fun registerInterface(
-        destination: String,
-        objectPath: String,
-        interfaceName: String,
-        propertiesProvider: () -> Map<PropertyName, Variant>
-    ): Resource {
-        val providerId = nextProviderId.getAndIncrement()
-        val key = ManagedObjectKey(destination, objectPath, interfaceName)
-        synchronized(this) {
-            interfaces
-                .getOrPut(key) { ManagedInterfaceEntry() }
-                .providers[providerId] = propertiesProvider
-        }
-        return ActionResource {
-            synchronized(this) {
-                val entry = interfaces[key] ?: return@synchronized
-                entry.providers.remove(providerId)
-                if (entry.providers.isEmpty()) {
-                    interfaces.remove(key)
-                }
-            }
-        }
-    }
-
-    fun managedObjectsFor(
-        managerPath: String,
-        destination: String
-    ): Map<ObjectPath, Map<InterfaceName, Map<PropertyName, Variant>>> = synchronized(this) {
-        val snapshot =
-            mutableMapOf<ObjectPath, MutableMap<InterfaceName, Map<PropertyName, Variant>>>()
-        interfaces.entries
-            .asSequence()
-            .filter { (key, _) ->
-                key.destination == destination &&
-                    key.objectPath.startsWith("$managerPath/") &&
-                    key.objectPath != managerPath
-            }
-            .forEach { (key, entry) ->
-                val objectPath = ObjectPath(key.objectPath)
-                val interfaceName = InterfaceName(key.interfaceName)
-                val properties = entry.providers.values
-                    .fold(emptyMap<PropertyName, Variant>()) { acc, provider ->
-                        acc + provider()
-                    }
-                snapshot.getOrPut(objectPath) { mutableMapOf() }[interfaceName] = properties
-            }
-        snapshot.mapValues { (_, interfacesForObject) -> interfacesForObject.toMap() }
-    }
-}
-
-internal object LocalObjectManagerRegistry {
-    private data class ManagerKey(val destination: String, val path: String)
-
-    private val managerPaths = mutableMapOf<String, Int>()
-    private val dispatchRegistrations = mutableMapOf<ManagerKey, Int>()
-
-    fun register(path: String, destination: String): Resource {
-        val key = ManagerKey(destination, path)
-        synchronized(this) {
-            managerPaths[path] = (managerPaths[path] ?: 0) + 1
-            val currentDispatchRefs = dispatchRegistrations[key] ?: 0
-            dispatchRegistrations[key] = currentDispatchRefs + 1
-            if (currentDispatchRefs == 0) {
-                JvmStaticDispatch.register(
-                    destination = destination,
-                    objectPath = path,
-                    interfaceName = OBJECT_MANAGER_INTERFACE_NAME,
-                    methodName = OBJECT_MANAGER_GET_MANAGED_OBJECTS,
-                    argCount = 0
-                ) {
-                    LocalManagedObjectsRegistry.managedObjectsFor(path, destination)
-                }
-            }
-        }
-        return ActionResource {
-            synchronized(this) {
-                val current = managerPaths[path] ?: return@synchronized
-                if (current <= 1) {
-                    managerPaths.remove(path)
-                } else {
-                    managerPaths[path] = current - 1
-                }
-
-                val currentDispatchRefs = dispatchRegistrations[key] ?: return@synchronized
-                if (currentDispatchRefs <= 1) {
-                    dispatchRegistrations.remove(key)
-                    JvmStaticDispatch.unregister(
-                        destination = destination,
-                        objectPath = path,
-                        interfaceName = OBJECT_MANAGER_INTERFACE_NAME,
-                        methodName = OBJECT_MANAGER_GET_MANAGED_OBJECTS,
-                        argCount = 0
-                    )
-                } else {
-                    dispatchRegistrations[key] = currentDispatchRefs - 1
-                }
-            }
-        }
-    }
-
-    fun resolveFor(objectPath: String): String? = synchronized(this) {
-        managerPaths.keys
-            .filter { manager ->
-                objectPath == manager || objectPath.startsWith("$manager/")
-            }
-            .maxByOrNull { it.length }
-    }
-}
-
-internal object LocalJvmServiceRegistry {
-    private val namesByUnique = mutableMapOf<String, MutableSet<String>>()
-
-    fun registerLocalUniqueName(uniqueName: String) {
-        if (uniqueName.isEmpty()) return
-        synchronized(this) {
-            namesByUnique.getOrPut(uniqueName) { mutableSetOf() } += uniqueName
-        }
-    }
-
-    fun unregisterLocalUniqueName(uniqueName: String) {
-        if (uniqueName.isEmpty()) return
-        synchronized(this) {
-            namesByUnique.remove(uniqueName)
-        }
-    }
-
-    fun addAlias(uniqueName: String, alias: String) {
-        if (uniqueName.isEmpty() || alias.isEmpty()) return
-        synchronized(this) {
-            namesByUnique.getOrPut(uniqueName) { mutableSetOf() }.apply {
-                add(uniqueName)
-                add(alias)
-            }
-        }
-    }
-
-    fun removeAlias(uniqueName: String, alias: String) {
-        if (uniqueName.isEmpty() || alias.isEmpty()) return
-        synchronized(this) {
-            val names = namesByUnique[uniqueName] ?: return@synchronized
-            if (alias != uniqueName) {
-                names.remove(alias)
-            }
-            if (names.isEmpty()) {
-                namesByUnique.remove(uniqueName)
-            }
-        }
-    }
-
-    fun resolveLocalUniqueName(destination: String): String? {
-        if (destination.isEmpty()) return null
-        synchronized(this) {
-            if (namesByUnique.containsKey(destination)) {
-                return destination
-            }
-            return namesByUnique.entries
-                .firstOrNull { (_, names) -> destination in names }
-                ?.key
-        }
-    }
-}
-
-private data class SenderCredentials(
-    val pid: Int?,
-    val uid: UInt?,
-    val gid: UInt?,
-    val supplementaryGids: List<UInt>?,
-    val selinuxContext: String?
-)
-
-private val localProcessCredentials: SenderCredentials by lazy {
-    val pid = runCatching { ProcessHandle.current().pid().toInt() }.getOrNull()
-    val unix = runCatching { UnixSystem() }.getOrNull()
-    val uid = unix?.uid?.toUInt()
-    val gid = unix?.gid?.toUInt()
-    val groups = unix?.groups?.map { it.toUInt() }
-    val selinuxContext = runCatching {
-        Files.readString(Paths.get("/proc/self/attr/current"))
-            .trim('\u0000', ' ')
-            .ifEmpty { null }
-    }.getOrNull()
-    SenderCredentials(
-        pid = pid,
-        uid = uid,
-        gid = gid,
-        supplementaryGids = groups,
-        selinuxContext = selinuxContext
-    )
-}
-
-private fun localProcessCredentialsOrNull(): SenderCredentials? = localProcessCredentials.takeIf {
-    it.pid != null ||
-        it.uid != null ||
-        it.gid != null ||
-        it.supplementaryGids != null ||
-        it.selinuxContext != null
-}
-
-private fun credentialUInt(value: Any?): UInt? = when (value) {
-    is UInt -> value
-    is Number -> value.toLong().takeIf { it >= 0L }?.toUInt()
-    is String -> value.toULongOrNull()?.toUInt()
-    else -> null
-}
-
-private fun credentialUIntList(value: Any?): List<UInt>? = when (value) {
-    is List<*> -> value.mapNotNull(::credentialUInt).takeIf { it.isNotEmpty() }
-    is Array<*> -> value.mapNotNull(::credentialUInt).takeIf { it.isNotEmpty() }
-    else -> null
-}
-
-private fun credentialBytes(value: Any?): ByteArray? = when (value) {
-    is ByteArray -> value
-    is List<*> -> value.mapNotNull { (it as? Number)?.toByte() }.toByteArray()
-        .takeIf { it.isNotEmpty() }
-    is Array<*> -> value.mapNotNull { (it as? Number)?.toByte() }.toByteArray()
-        .takeIf { it.isNotEmpty() }
-
-    else -> null
-}
-
-private fun credentialSelinuxContext(value: Any?): String? = when (value) {
-    is String -> value.trim().ifEmpty { null }
-    else -> credentialBytes(value)
-        ?.toString(Charsets.UTF_8)
-        ?.trim('\u0000', ' ')
-        ?.ifEmpty { null }
-}
-
-private fun firstCredentialValue(
-    credentials: Map<String, org.freedesktop.dbus.types.Variant<*>>,
-    keys: List<String>
-): Any? = keys.firstNotNullOfOrNull { key -> credentials[key]?.value }
-
-private fun resolveSenderCredentials(
-    connection: AbstractConnection?,
-    sender: String?
-): SenderCredentials? {
-    val senderName = sender?.takeIf { it.isNotBlank() } ?: return null
-    val busConnection = connection as? DBusConnection
-    val localSender = LocalJvmServiceRegistry.resolveLocalUniqueName(senderName) != null ||
-        busConnection?.uniqueName == senderName
-    val resolvedBusConnection = busConnection
-        ?: return if (localSender) localProcessCredentialsOrNull() else null
-    val dbus = runCatching {
-        resolvedBusConnection.dynamicProxy(
-            "org.freedesktop.DBus",
-            "/org/freedesktop/DBus",
-            DBus::class.java
-        )
-    }.getOrNull() ?: return if (localSender) localProcessCredentialsOrNull() else null
-
-    val credentials = runCatching {
-        dbus.GetConnectionCredentials(senderName)
-    }.getOrNull().orEmpty()
-
-    val pid = credentialUInt(
-        firstCredentialValue(credentials, listOf("ProcessID", "UnixProcessID"))
-    )?.toInt() ?: runCatching {
-        dbus.GetConnectionUnixProcessID(senderName).toInt()
-    }.getOrNull()
-
-    val uid = credentialUInt(
-        firstCredentialValue(credentials, listOf("UnixUserID", "UserID"))
-    ) ?: runCatching {
-        dbus.GetConnectionUnixUser(senderName).toLong().toUInt()
-    }.getOrNull()
-
-    val supplementaryGids = credentialUIntList(
-        firstCredentialValue(credentials, listOf("UnixGroupIDs", "GroupIDs"))
-    )
-    val gid = credentialUInt(
-        firstCredentialValue(credentials, listOf("UnixGroupID", "GroupID"))
-    ) ?: supplementaryGids?.firstOrNull()
-
-    val selinuxContext = credentialSelinuxContext(
-        firstCredentialValue(credentials, listOf("LinuxSecurityLabel", "SELinuxContext"))
-    ) ?: runCatching {
-        credentialSelinuxContext(dbus.GetConnectionSELinuxSecurityContext(senderName))
-    }.getOrNull()
-
-    if (
-        pid == null &&
-        uid == null &&
-        gid == null &&
-        supplementaryGids == null &&
-        selinuxContext == null
-    ) {
-        return if (localSender) localProcessCredentialsOrNull() else null
-    }
-    return SenderCredentials(
-        pid = pid,
-        uid = uid,
-        gid = gid,
-        supplementaryGids = supplementaryGids,
-        selinuxContext = selinuxContext
-    )
-}
-
-private fun Message.Metadata.withSenderCredentials(
-    connection: AbstractConnection?,
-    sender: String?
-): Message.Metadata {
-    val credentials = resolveSenderCredentials(connection, sender) ?: return this
-    return copy(
-        credsPid = credentials.pid,
-        credsUid = credentials.uid,
-        credsEuid = credentials.uid,
-        credsGid = credentials.gid,
-        credsEgid = credentials.gid,
-        credsSupplementaryGids = credentials.supplementaryGids,
-        selinuxContext = credentials.selinuxContext
-    )
-}
-
-private fun parseMatchFilter(match: String): MatchFilter {
-    val values = mutableMapOf<String, String>()
-    match.split(",")
-        .map { it.trim() }
-        .filter { it.isNotEmpty() }
-        .forEach { token ->
-            val parts = token.split("=", limit = 2)
-            if (parts.size == 2) {
-                val key = parts[0].trim()
-                val value = parts[1].trim().trim('\'').trim('"')
-                values[key] = value
-            }
-        }
-    return MatchFilter(
-        sender = values["sender"],
-        path = values["path"],
-        interfaceName = values["interface"],
-        member = values["member"]
-    )
-}
-
-private fun DBusSignal.matches(filter: MatchFilter): Boolean {
-    val senderOk = filter.sender == null || source == filter.sender
-    val pathOk = filter.path == null || path == filter.path
-    val interfaceOk = filter.interfaceName == null || `interface` == filter.interfaceName
-    val memberOk = filter.member == null || name == filter.member
-    return senderOk && pathOk && interfaceOk && memberOk
-}
-
-private fun DBusSignal.toSdbusMessage(connection: AbstractConnection?): Signal = signalFromMetadata(
-    Message.Metadata(
-        interfaceName = `interface`,
-        memberName = name,
-        sender = source,
-        path = path,
-        destination = destination,
-        valid = true,
-        empty = false
-    ).withSenderCredentials(connection, source)
-).also { signal ->
-    val values = runCatching { parameters.toList() }.getOrElse { emptyList() }
-    fromJavaWireValues(values, runCatching { sig }.getOrNull())
-        .forEach { signal.payload.add(it) }
-}
-
-private fun fromJavaSignalValue(value: Any?): Any? = when (value) {
-    is UInt16 -> value.toInt().toUShort()
-    is UInt32 -> value.toLong().toUInt()
-    is UInt64 -> value.value().toString().toULong()
-    is org.freedesktop.dbus.DBusPath -> ObjectPath(value.path)
-    is FileDescriptor -> UnixFd(value.intFileDescriptor)
-    is org.freedesktop.dbus.types.Variant<*> -> fromJavaVariantValue(value)
-    is List<*> -> value.map(::fromJavaSignalValue)
-    is Array<*> -> value.map(::fromJavaSignalValue).toTypedArray()
-    is Map<*, *> -> value.entries.associate { (k, v) ->
-        fromJavaSignalValue(k) to fromJavaSignalValue(v)
-    }
-
-    else -> value
-}
-
-private fun toJavaSignalValue(value: Any?): Any? = when (value) {
-    is UByte -> value.toByte()
-    is UShort -> UInt16(value.toInt())
-    is UInt -> UInt32(value.toLong())
-    is ULong -> UInt64(value.toString())
-    is BusName -> value.value
-    is InterfaceName -> value.value
-    is MemberName -> value.value
-    is UnixFd -> FileDescriptor(value.fd)
-    is com.monkopedia.sdbus.ObjectPath -> value.value
-    is Signature -> value.value
-    is Message.JvmVariantPayload -> toJavaVariantValue(value.signature, value.value)
-    // dbus-java marshals struct values from positional Object[] contents (see its
-    // Message.appendOne STRUCT1 case), so the decomposed struct payload converts its
-    // fields in order (issue #71).
-    is Message.JvmStructPayload -> value.fields.map(::toJavaSignalValue).toTypedArray()
-    is Variant -> extractVariantPayload(value).let { payload ->
-        toJavaVariantValue(payload.signature, payload.value)
-    }
-    is List<*> -> value.map(::toJavaSignalValue)
-    is Array<*> -> value.map(::toJavaSignalValue).toTypedArray()
-    is Map<*, *> -> value.entries.associate { (k, v) ->
-        toJavaSignalValue(k) to toJavaSignalValue(v)
-    }
-
-    else -> value
-}
-
-private fun fromJavaVariantValue(value: org.freedesktop.dbus.types.Variant<*>): Variant {
-    val signature = value.sig
-        ?: inferSignalSignature(value.value)
-        ?: throw createError(
-            -1,
-            "fromJavaVariantValue failed: unsupported payload type ${value.value?.javaClass}"
-        )
-    val message = PlainMessage.createPlainMessage()
-    message.payload.add(
-        Message.JvmVariantPayload(
-            signature,
-            fromJavaWireValue(value.value, signature)
-        )
-    )
-    return Variant().apply { deserializeFrom(message) }
-}
-
-private fun extractVariantPayload(variant: Variant): Message.JvmVariantPayload {
-    val signature = variant.peekValueType()
-        ?: throw createError(-1, "Cannot convert empty variant to JVM D-Bus payload")
-    val message = PlainMessage.createPlainMessage()
-    variant.serializeTo(message)
-    message.rewind(false)
-    message.enterVariant(signature)
-    val value = message.nextDeserializedValue("extractVariantPayload")
-    message.exitVariant()
-    return Message.JvmVariantPayload(signature, value)
-}
-
-private fun toJavaVariantValue(
-    signature: String,
-    value: Any?
-): org.freedesktop.dbus.types.Variant<Any?> = org.freedesktop.dbus.types.Variant(
-    toJavaSignalValue(value),
-    signature
-)
-
-// Splits a D-Bus signature into its top-level single complete types
-// (e.g. "sa{sv}(is)" -> ["s", "a{sv}", "(is)"]).
-private fun splitTopLevelTypes(signature: String): List<String> {
-    fun parseOne(index: Int): Int {
-        if (index >= signature.length) return index
-        return when (signature[index]) {
-            'a' -> parseOne(index + 1)
-            '(' -> {
-                var i = index + 1
-                while (i < signature.length && signature[i] != ')') {
-                    i = parseOne(i)
-                }
-                (i + 1).coerceAtMost(signature.length)
-            }
-            '{' -> {
-                var i = index + 1
-                while (i < signature.length && signature[i] != '}') {
-                    i = parseOne(i)
-                }
-                (i + 1).coerceAtMost(signature.length)
-            }
-            else -> index + 1
-        }
-    }
-
-    val types = mutableListOf<String>()
-    var index = 0
-    while (index < signature.length) {
-        val next = parseOne(index)
-        if (next <= index) break
-        types += signature.substring(index, next)
-        index = next
-    }
-    return types
-}
-
-// Counts the number of top-level D-Bus types in a signature (e.g. "sa{sv}" -> 2),
-// used to check that a declared body signature lines up with the payload it describes.
-private fun countTopLevelTypes(signature: String?): Int =
-    if (signature.isNullOrEmpty()) 0 else splitTopLevelTypes(signature).size
-
-// Signature-aware conversion of a dbus-java message body into the JVM payload model. The
-// wire signature is authoritative: it is the only way to tell a struct value (extracted by
-// dbus-java as positional Object[] contents) apart from an array, so structs become
-// Message.JvmStructPayload carrying their exact signature (issue #71). Values whose
-// signature is unavailable fall back to the value-shape-based conversion.
-private fun fromJavaWireValues(values: List<Any?>, signature: String?): List<Any?> {
-    val types = signature?.let(::splitTopLevelTypes).orEmpty()
-    return values.mapIndexed { index, value -> fromJavaWireValue(value, types.getOrNull(index)) }
-}
-
-private fun fromJavaWireValue(value: Any?, signature: String?): Any? {
-    if (value == null || signature.isNullOrEmpty()) return fromJavaSignalValue(value)
-    return when {
-        signature.startsWith("(") && signature.endsWith(")") -> {
-            val fields = wireValueAsList(value) ?: return fromJavaSignalValue(value)
-            val fieldTypes = splitTopLevelTypes(signature.substring(1, signature.length - 1))
-            Message.JvmStructPayload(
-                signature,
-                fields.mapIndexed { index, field ->
-                    fromJavaWireValue(field, fieldTypes.getOrNull(index))
-                }
-            )
-        }
-
-        signature.startsWith("a{") && signature.endsWith("}") -> {
-            val map = value as? Map<*, *> ?: return fromJavaSignalValue(value)
-            val entryTypes = splitTopLevelTypes(signature.substring(2, signature.length - 1))
-            map.entries.associate { (key, entryValue) ->
-                fromJavaWireValue(key, entryTypes.getOrNull(0)) to
-                    fromJavaWireValue(entryValue, entryTypes.getOrNull(1))
-            }
-        }
-
-        signature.startsWith("a") -> {
-            val items = wireValueAsList(value) ?: return fromJavaSignalValue(value)
-            val elementType = signature.substring(1)
-            items.map { fromJavaWireValue(it, elementType) }
-        }
-
-        // Primitives and variants: variants carry their own signature on the dbus-java side,
-        // which fromJavaVariantValue (via fromJavaSignalValue) follows recursively.
-        else -> fromJavaSignalValue(value)
-    }
-}
-
-private fun wireValueAsList(value: Any?): List<Any?>? = when (value) {
-    is List<*> -> value
-    is Array<*> -> value.toList()
-    is ByteArray -> value.toList()
-    is ShortArray -> value.toList()
-    is IntArray -> value.toList()
-    is LongArray -> value.toList()
-    is BooleanArray -> value.toList()
-    is FloatArray -> value.toList()
-    is DoubleArray -> value.toList()
-    else -> null
-}
-
-// Wire signature for an outgoing body. Prefer the declared signature accumulated from
-// serializer descriptors (correct even for empty collections); only fall back to value-based
-// inference when no usable declared signature is available — e.g. a body assembled through the
-// raw append() API rather than serialize(). The top-level type count guards against a declared
-// signature that doesn't line up with the payload.
-private fun bodySignature(
-    payload: List<Any?>,
-    declaredSignature: String?,
-    errorFor: (Any?) -> String
-): String {
-    if (!declaredSignature.isNullOrEmpty() &&
-        countTopLevelTypes(declaredSignature) == payload.size
-    ) {
-        return declaredSignature
-    }
-    return payload.joinToString(separator = "") { value ->
-        inferSignalSignature(value) ?: throw createError(-1, errorFor(value))
-    }
-}
-
-private fun inferSignalSignature(value: Any?): String? = when (value) {
-    null -> null
-    is Message.JvmVariantPayload -> "v"
-    is Message.JvmStructPayload -> value.signature
-    is Variant -> "v"
-    is Boolean -> "b"
-    is Byte, is UByte -> "y"
-    is Short -> "n"
-    is UShort -> "q"
-    is Int -> "i"
-    is UInt -> "u"
-    is Long -> "x"
-    is ULong -> "t"
-    is Float, is Double -> "d"
-    is String -> "s"
-    is BusName,
-    is InterfaceName,
-    is MemberName -> "s"
-    is com.monkopedia.sdbus.ObjectPath -> "o"
-    is Signature -> "g"
-    is UnixFd -> "h"
-    is List<*> -> {
-        val elementSig = value.firstNotNullOfOrNull(::inferSignalSignature) ?: return "a"
-        "a$elementSig"
-    }
-
-    is Array<*> -> {
-        val elementSig = value.firstNotNullOfOrNull(::inferSignalSignature) ?: return "a"
-        "a$elementSig"
-    }
-
-    is Map<*, *> -> {
-        val first = value.entries.firstOrNull() ?: return "a{}"
-        val keySig = inferSignalSignature(first.key) ?: return "a{}"
-        val valueSig = inferSignalSignature(first.value) ?: return "a{}"
-        "a{$keySig$valueSig}"
-    }
-
-    else -> null
-}
-
-// A remote error reply carries the D-Bus error name in its ERROR_NAME header (surfaced by
-// dbus-java's Message.getName() for Error messages) and the human-readable message as the
-// first body argument. Both must be preserved verbatim like the native backend does (sd-bus
-// copies them straight into sd_bus_error) instead of being squeezed through the errno-based
-// createError mapping, which made every foreign error surface as AccessDenied (issue #72).
-private fun org.freedesktop.dbus.messages.Error.toSdbusError(): com.monkopedia.sdbus.Error {
-    val wireName = runCatching { name }.getOrNull()?.takeIf { it.isNotBlank() }
-        ?: return createError(-1, "Remote method call failed")
-    val wireMessage = runCatching { parameters.firstOrNull() as? String }.getOrNull().orEmpty()
-    return com.monkopedia.sdbus.Error(wireName, wireMessage)
-}
-
-private fun AbstractConnection.uniqueNameOrNull(): String? = (this as? DBusConnection)?.uniqueName
-
-private class PureJavaDbusConnection(internal val javaConnection: AbstractConnection) :
-    JvmDbusConnection {
-    private var timeout: Duration = Duration.ZERO
-    private val localUniqueName: String = javaConnection.uniqueNameOrNull().orEmpty()
-    private val released = AtomicBoolean(false)
-
-    fun isReleased(): Boolean = released.get()
-
-    init {
-        LocalJvmServiceRegistry.registerLocalUniqueName(localUniqueName)
-    }
-
-    override fun startEventLoop(): Unit = Unit
-
-    override suspend fun stopEventLoop(): Unit = Unit
-
-    override fun currentlyProcessedMessage(): Message =
-        JvmCurrentMessageContext.current() ?: com.monkopedia.sdbus.PlainMessage.createPlainMessage()
-
-    override fun setMethodCallTimeout(timeout: Duration): Unit = run {
-        this.timeout = timeout
-    }
-
-    override fun getMethodCallTimeout(): Duration = timeout
-
-    override fun addObjectManager(objectPath: ObjectPath): Resource =
-        LocalObjectManagerRegistry.register(objectPath.value, localUniqueName)
-
-    override fun uniqueName(): BusName = BusName(localUniqueName.ifEmpty { ":jvm-direct" })
-
-    override fun requestName(name: ServiceName): Unit = run {
-        val busConnection = javaConnection as? DBusConnection ?: return
-        runCatching {
-            busConnection.requestBusName(name.value)
-        }.getOrElse {
-            throw createError(-1, "requestBusName failed: ${it.message}")
-        }
-        LocalJvmServiceRegistry.addAlias(localUniqueName, name.value)
-    }
-
-    override fun releaseName(name: ServiceName): Unit = run {
-        val busConnection = javaConnection as? DBusConnection ?: return
-        runCatching {
-            busConnection.releaseBusName(name.value)
-        }.getOrElse {
-            throw createError(-1, "releaseBusName failed: ${it.message}")
-        }
-        LocalJvmServiceRegistry.removeAlias(localUniqueName, name.value)
-    }
-
-    override fun addMatch(match: String, callback: MessageHandler): Resource {
-        val filter = parseMatchFilter(match)
-        val matchRule = DBusMatchRule(
-            "signal",
-            filter.interfaceName,
-            filter.member,
-            filter.path
-        )
-        val closeable = runCatching {
-            val handler = DBusSigHandler<DBusSignal> { signal ->
-                if (signal.matches(filter)) {
-                    val message = signal.toSdbusMessage(javaConnection)
-                    JvmCurrentMessageContext.withMessage(message) {
-                        callback(message)
-                    }
-                }
-            }
-            if (javaConnection is DBusConnection) {
-                javaConnection.addGenericSigHandler(matchRule, handler)
-            } else {
-                addGenericSigHandler(javaConnection, matchRule, handler)
-            }
-        }.getOrElse {
-            throw createError(-1, "addMatch failed: ${it.message}")
-        }
-        return CloseableResource(closeable)
-    }
-
-    override fun addMatchAsync(
-        match: String,
-        callback: MessageHandler,
-        installCallback: MessageHandler
-    ): Resource {
-        val resource = addMatch(match, callback)
-        installCallback(signalFromMetadata(Message.Metadata(valid = true, empty = true)))
-        return resource
-    }
-
-    override fun release(): Unit = run {
-        if (!released.compareAndSet(false, true)) return@run
-        LocalJvmServiceRegistry.unregisterLocalUniqueName(localUniqueName)
-        javaConnection.disconnect()
-    }
-}
-
-private class PureJavaDbusProxy(
-    private val connection: AbstractConnection?,
-    private val destination: ServiceName,
-    private val objectPath: ObjectPath,
-    private val callerName: String?,
-    private val isConnectionReleased: () -> Boolean = { false },
-    private val defaultTimeoutMicros: () -> ULong = { 0uL }
-) : JvmDbusProxy {
-    private fun invalidReply(path: String, interfaceName: String, methodName: String): MethodReply =
-        methodReplyFrom(
-            Message.Metadata(
-                interfaceName = interfaceName,
-                memberName = methodName,
-                sender = destination.value,
-                path = path,
-                destination = connection?.uniqueNameOrNull() ?: destination.value,
-                valid = false,
-                empty = true
-            ),
-            emptyList()
-        )
-
-    override fun currentlyProcessedMessage(): Message =
-        JvmCurrentMessageContext.current() ?: signalFromMetadata(Message.Metadata())
-
-    // A call made after the proxy's connection has been released must fail like it does on
-    // the native backend instead of being silently serviced via the in-process
-    // static-dispatch short circuit. Native parity: ConnectionImpl guards every
-    // post-release operation with require(!released) { "Connection has already been
-    // released" }, and a connection that dropped without an explicit release surfaces
-    // -ENOTCONN from sd_bus_call (Error("System.Error.ENOTCONN", ...)).
-    private fun requireConnectionUsable() {
-        require(!isConnectionReleased()) { "Connection has already been released" }
-        val realConnection = connection ?: return
-        if (!realConnection.isConnected) {
-            // ENOTCONN, mirroring the native error name System.Error.ENOTCONN.
-            throw createError(107, "callMethod failed: connection is not connected")
-        }
-    }
-
-    override fun createMethodCall(
-        interfaceName: InterfaceName,
-        methodName: MethodName
-    ): MethodCall = com.monkopedia.sdbus.MethodCall().also {
-        it.metadata = Message.Metadata(
-            interfaceName = interfaceName.value,
-            memberName = methodName.value,
-            destination = destination.value,
-            path = objectPath.value,
-            valid = true,
-            empty = true
-        )
-    }
-
-    // A call without a timeout argument carries the same "no explicit timeout" meaning as the
-    // 0 sentinel (Duration.ZERO at the API surface), so it funnels through the same resolution
-    // and picks up the connection-level methodCallTimeout default when one is set (issue #80).
-    override fun callMethod(message: MethodCall): MethodReply = callMethod(message, 0uL)
-
-    private fun callStaticDispatch(
-        message: MethodCall,
-        interfaceName: String,
-        methodName: String,
-        path: String,
-        dispatchDestination: String
-    ): MethodReply {
-        val inboundSender = callerName ?: connection?.uniqueNameOrNull()
-        val inbound = MethodCall().also {
-            it.metadata = Message.Metadata(
-                interfaceName = interfaceName,
-                memberName = methodName,
-                sender = inboundSender,
-                destination = destination.value,
-                path = path,
-                valid = true,
-                empty = message.payload.isEmpty()
-            ).withSenderCredentials(connection, inboundSender)
-            it.payload.addAll(message.payload)
-        }
-        val result = runCatching {
-            JvmCurrentMessageContext.withMessage(inbound) {
-                JvmStaticDispatch.invokeOrNull(
-                    objectPath = path,
-                    interfaceName = interfaceName,
-                    methodName = methodName,
-                    args = message.payload,
-                    destination = dispatchDestination
-                )
-            }
-        }.getOrElse {
-            throw createError(-1, "callMethod failed: ${it.message}")
-        } ?: throw createError(
-            -1,
-            "callMethod failed: no static binding for $path:$interfaceName.$methodName/${message.payload.size}"
-        )
-        val values = when (result) {
-            Unit -> emptyList()
-            is JvmStaticDispatch.DispatchResult -> {
-                result.reply.error?.let { throw it }
-                if (!result.reply.isValid) {
-                    throw createError(-1, "callMethod failed: method returned an invalid reply")
-                }
-                result.reply.payload.toList()
-            }
-            else -> listOf(result)
-        }
-        return methodReplyFrom(
-            Message.Metadata(
-                interfaceName = interfaceName,
-                memberName = methodName,
-                sender = destination.value,
-                path = path,
-                destination = connection?.uniqueNameOrNull() ?: destination.value,
-                valid = true,
-                empty = values.isEmpty()
-            ),
-            values
-        )
-    }
-
-    override fun callMethod(message: MethodCall, timeout: ULong): MethodReply {
-        requireConnectionUsable()
-        // Native parity (issue #80): an explicit per-call timeout always wins; the
-        // connection-level methodCallTimeout only fills the unset (0) case -- exactly how
-        // sd_bus_call resolves a 0 timeout via sd_bus_get_method_call_timeout. If neither is
-        // set, the effective timeout stays 0 and the pre-existing default behavior applies.
-        val effectiveTimeout = if (timeout == 0uL) defaultTimeoutMicros() else timeout
-        val interfaceName = message.interfaceName?.value
-            ?: throw createError(-1, "callMethod failed: missing interface name")
-        val methodName = message.memberName?.value
-            ?: throw createError(-1, "callMethod failed: missing method name")
-        val path = message.path?.value ?: objectPath.value
-        val realConnection = connection
-        if (realConnection != null) {
-            val localDestination = resolveLocalDispatchDestination(realConnection)
-            val dispatchDestination = localDestination ?: destination.value
-            if (
-                JvmStaticDispatch.hasHandler(
-                    objectPath = path,
-                    interfaceName = interfaceName,
-                    methodName = methodName,
-                    args = message.payload,
-                    destination = dispatchDestination
-                )
-            ) {
-                return if (effectiveTimeout == 0uL) {
-                    callStaticDispatch(
-                        message = message,
-                        interfaceName = interfaceName,
-                        methodName = methodName,
-                        path = path,
-                        dispatchDestination = dispatchDestination
-                    )
-                } else {
-                    callWithTimeout(
-                        message = message,
-                        interfaceName = interfaceName,
-                        methodName = methodName,
-                        path = path,
-                        timeout = effectiveTimeout,
-                        dispatchDestination = dispatchDestination
-                    )
-                }
-            }
-            return callRemoteMethod(
-                connection = realConnection,
-                interfaceName = interfaceName,
-                methodName = methodName,
-                path = path,
-                payload = message.payload.toList(),
-                declaredSignature = message.declaredBodySignature.toString(),
-                timeout = effectiveTimeout.takeUnless { it == 0uL }
-            )
-        }
-        if (effectiveTimeout == 0uL) {
-            return callStaticDispatch(
-                message = message,
-                interfaceName = interfaceName,
-                methodName = methodName,
-                path = path,
-                dispatchDestination = destination.value
-            )
-        }
-        return callWithTimeout(
-            message = message,
-            interfaceName = interfaceName,
-            methodName = methodName,
-            path = path,
-            timeout = effectiveTimeout,
-            dispatchDestination = destination.value
-        )
-    }
-
-    private fun callWithTimeout(
-        message: MethodCall,
-        interfaceName: String,
-        methodName: String,
-        path: String,
-        timeout: ULong,
-        dispatchDestination: String
-    ): MethodReply {
-        val result = AtomicReference<MethodReply?>()
-        val failure = AtomicReference<Throwable?>()
-        val done = java.util.concurrent.CountDownLatch(1)
-        thread(start = true, isDaemon = true, name = "sdbus-jvm-sync-timeout") {
-            runCatching {
-                callStaticDispatch(
-                    message = message,
-                    interfaceName = interfaceName,
-                    methodName = methodName,
-                    path = path,
-                    dispatchDestination = dispatchDestination
-                )
-            }
-                .onSuccess { result.set(it) }
-                .onFailure { failure.set(it) }
-            done.countDown()
-        }
-        val timeoutMillis = ((timeout + 999uL) / 1000uL).toLong().coerceAtLeast(1L)
-        if (!done.await(timeoutMillis, java.util.concurrent.TimeUnit.MILLISECONDS)) {
-            throw createError(-1, "Method call timed out")
-        }
-        failure.get()?.let { throw it }
-        return result.get() ?: throw createError(-1, "callMethod failed: missing result")
-    }
-
-    private fun resolveLocalDispatchDestination(realConnection: AbstractConnection): String? {
-        val destinationName = destination.value
-        val localUniqueName = realConnection.uniqueNameOrNull().orEmpty()
-        if (localUniqueName.isEmpty()) return null
-        if (destinationName == localUniqueName) return localUniqueName
-
-        val localRegistryOwner = LocalJvmServiceRegistry.resolveLocalUniqueName(destinationName)
-        if (localRegistryOwner == localUniqueName) return localUniqueName
-
-        val localBusOwner = (realConnection as? DBusConnection)?.let { connection ->
-            if (destinationName.isEmpty()) return@let null
-            runCatching {
-                connection.getDBusOwnerName(destinationName)
-            }.getOrNull()
-        }
-        return localUniqueName.takeIf { it == localBusOwner }
-    }
-
-    private fun callRemoteMethod(
-        connection: AbstractConnection,
-        interfaceName: String,
-        methodName: String,
-        path: String,
-        payload: List<Any?>,
-        declaredSignature: String?,
-        timeout: ULong?
-    ): MethodReply {
-        val signature = if (payload.isEmpty()) {
-            null
-        } else {
-            bodySignature(payload, declaredSignature) { value ->
-                "callMethod failed: unsupported argument type ${
-                    value?.let { typed -> typed::class.simpleName } ?: "null"
-                }"
-            }
-        }
-        val args = payload.map(::toJavaSignalValue).toTypedArray()
-        val destinationName = destination.value.ifEmpty { null }
-        val rawCall = runCatching {
-            connection.messageFactory.createMethodCall(
-                null,
-                destinationName,
-                path,
-                interfaceName,
-                methodName,
-                0,
-                signature,
-                *args
-            )
-        }.getOrElse {
-            throw createError(-1, "callMethod failed: ${it.message}")
-        }
-        runCatching {
-            connection.sendMessage(rawCall)
-        }.getOrElse {
-            throw createError(-1, "callMethod failed: ${it.message}")
-        }
-        val rawReply = runCatching {
-            if (timeout == null) {
-                rawCall.reply
-            } else {
-                rawCall.getReply(((timeout + 999uL) / 1000uL).toLong().coerceAtLeast(1L))
-            }
-        }.getOrElse {
-            throw createError(-1, "callMethod failed: ${it.message}")
-        } ?: throw createError(-1, "Method call timed out")
-
-        if (rawReply is org.freedesktop.dbus.messages.Error) {
-            throw rawReply.toSdbusError()
-        }
-        val values = runCatching {
-            fromJavaWireValues(
-                rawReply.parameters.toList(),
-                runCatching { rawReply.sig }.getOrNull()
-            )
-        }.getOrElse {
-            throw createError(-1, "callMethod failed: ${it.message}")
-        }
-        val sender = rawReply.source ?: destination.value
-        return methodReplyFrom(
-            Message.Metadata(
-                interfaceName = interfaceName,
-                memberName = methodName,
-                sender = sender,
-                path = rawReply.path ?: path,
-                destination = rawReply.destination,
-                valid = true,
-                empty = values.isEmpty()
-            ).withSenderCredentials(connection, sender),
-            values
-        )
-    }
-
-    override fun callMethodAsync(
-        message: MethodCall,
-        asyncReplyCallback: com.monkopedia.sdbus.AsyncReplyHandler
-    ): PendingAsyncCall = callMethodAsync(message, asyncReplyCallback, 0u)
-
-    override fun callMethodAsync(
-        message: MethodCall,
-        asyncReplyCallback: com.monkopedia.sdbus.AsyncReplyHandler,
-        timeout: ULong
-    ): PendingAsyncCall {
-        val cancelled = AtomicBoolean(false)
-        val pending = AtomicBoolean(true)
-        val interfaceName = message.interfaceName?.value.orEmpty()
-        val methodName = message.memberName?.value.orEmpty()
-        val path = message.path?.value ?: objectPath.value
-
-        thread(
-            start = true,
-            isDaemon = true,
-            name = "sdbus-jvm-call-$interfaceName.$methodName"
-        ) {
-            val outcome = runCatching {
-                if (timeout == 0uL) callMethod(message) else callMethod(message, timeout)
-            }
-            pending.set(false)
-            if (cancelled.get()) return@thread
-            outcome.fold(
-                onSuccess = { asyncReplyCallback(it, null) },
-                onFailure = {
-                    val error = it as? com.monkopedia.sdbus.Error
-                        ?: createError(-1, it.message ?: "JVM async call failed")
-                    asyncReplyCallback(invalidReply(path, interfaceName, methodName), error)
-                }
-            )
-        }
-
-        return PendingAsyncCall(
-            cancelAction = { cancelled.set(true) },
-            isPendingAction = { pending.get() }
-        )
-    }
-
-    override suspend fun callMethodAsync(message: MethodCall): MethodReply = callMethod(message)
-
-    override suspend fun callMethodAsync(message: MethodCall, timeout: ULong): MethodReply =
-        if (timeout == 0uL) callMethod(message) else callMethod(message, timeout)
-
-    override fun registerSignalHandler(
-        interfaceName: InterfaceName,
-        signalName: SignalName,
-        signalHandler: SignalHandler
-    ): Resource {
-        val realConnection = connection
-        if (realConnection == null) {
-            return LocalJvmSignalBus.register(
-                destination = destination.value,
-                path = objectPath.value,
-                interfaceName = interfaceName.value,
-                member = signalName.value,
-                signalHandler = signalHandler
-            )
-        }
-        val resolvedDestination = (realConnection as? DBusConnection)?.let { connection ->
-            runCatching {
-                connection.getDBusOwnerName(destination.value)
-            }.getOrNull()
-        }
-        val matchRule = DBusMatchRule(
-            "signal",
-            interfaceName.value,
-            signalName.value,
-            objectPath.value
-        )
-        val closeable = runCatching {
-            val callback = DBusSigHandler<DBusSignal> { signal ->
-                val matchesDestination = destination.value.isEmpty() ||
-                    signal.source == destination.value ||
-                    (
-                        resolvedDestination?.isNotEmpty() == true &&
-                            signal.source == resolvedDestination
-                        )
-                val matchesPath = signal.path == objectPath.value
-                val matchesInterface = signal.`interface` == interfaceName.value
-                val matchesMember = signal.name == signalName.value
-                if (matchesDestination && matchesPath && matchesInterface && matchesMember) {
-                    val message = signal.toSdbusMessage(realConnection)
-                    JvmCurrentMessageContext.withMessage(message) {
-                        signalHandler(message)
-                    }
-                }
-            }
-            if (realConnection is DBusConnection) {
-                realConnection.addGenericSigHandler(matchRule, callback)
-            } else {
-                addGenericSigHandler(realConnection, matchRule, callback)
-            }
-        }.getOrElse {
-            throw createError(-1, "registerSignalHandler failed: ${it.message}")
-        }
-        return CloseableResource(closeable)
-    }
-
-    override fun release(): Unit = Unit
-}
-
-private fun addGenericSigHandler(
-    connection: AbstractConnection,
-    matchRule: DBusMatchRule,
-    callback: DBusSigHandler<DBusSignal>
-): AutoCloseable {
-    val method = AbstractConnection::class.java.getDeclaredMethod(
-        "addGenericSigHandler",
-        DBusMatchRule::class.java,
-        DBusSigHandler::class.java
-    ).apply { isAccessible = true }
-    return method.invoke(connection, matchRule, callback) as AutoCloseable
-}
-
-/**
- * Sends a signal OVER THE WIRE on a self-owned connection (epic #93 phase 3b). Supplied to
- * [PureJavaDbusObject] by the wire backend in place of in-process [LocalJvmSignalBus] delivery, so
- * an object served on the owned connection emits a real D-Bus SIGNAL that the bus routes to every
- * subscriber — external processes AND same-JVM connections (each of which has its own socket +
- * `AddMatch`). The sender header is left unset: the bus stamps the authoritative sender (our unique
- * name) and attaches the sender's credentials, which is what lets a receiver resolve them.
- */
-internal fun interface WireSignalEmitter {
-    fun emitSignalOverWire(
-        path: String,
-        interfaceName: String,
-        member: String,
-        signature: String?,
-        payload: List<Any?>
-    )
-}
-
-internal class PureJavaDbusObject(
-    private val objectPath: ObjectPath,
-    private val javaConnection: AbstractConnection?,
-    private val senderName: String?,
-    // When non-null (the wire backend), signals this object emits go over the wire instead of
-    // through the in-process [LocalJvmSignalBus]. The two are mutually exclusive on a given
-    // backend: dbus-java sends through its own connection (javaConnection != null); the wire
-    // backend has javaConnection == null and routes here; the in-process stub path keeps using
-    // [LocalJvmSignalBus]. This keeps signal delivery uniform with the native backend (always over
-    // the wire) for the owned connection. See WireDbusBackend.createObject.
-    private val wireSignalEmitter: WireSignalEmitter? = null
-) : JvmDbusObject {
-
-    // Routes a signal whose owning connection is NOT dbus-java (javaConnection == null): over the
-    // wire when the wire backend supplied an emitter, otherwise through the in-process bus (the
-    // stub/no-connection path). The wire path needs the marshalling [signature]; the in-process
-    // path ignores it.
-    private fun emitLocalOrWire(
-        sender: String?,
-        path: String,
-        interfaceName: String,
-        member: String,
-        signature: String?,
-        payload: List<Any?>
-    ) {
-        val emitter = wireSignalEmitter
-        if (emitter != null) {
-            emitter.emitSignalOverWire(path, interfaceName, member, signature, payload)
-        } else {
-            LocalJvmSignalBus.emit(sender, path, interfaceName, member, payload)
-        }
-    }
-    private val registrations = mutableListOf<Resource>()
-    private val registeredInterfaces = mutableMapOf<String, Int>()
-    private val asyncReplyTimeoutMs = 10_000L
-    private val propertiesInterfaceName = "org.freedesktop.DBus.Properties"
-    private val propertiesChangedMemberName = "PropertiesChanged"
-    private val propertiesChangedSignature = "sa{sv}as"
-    private val objectManagerInterfaceName = OBJECT_MANAGER_INTERFACE_NAME
-    private val interfacesAddedMemberName = "InterfacesAdded"
-    private val interfacesRemovedMemberName = "InterfacesRemoved"
-    private val interfacesAddedSignature = "oa{sa{sv}}"
-    private val interfacesRemovedSignature = "oas"
-    private val propertiesByInterface =
-        mutableMapOf<String, MutableMap<String, PropertyVTableItem>>()
-    private var propertiesDispatchRegistered = false
-    private val dispatchDestination = senderName ?: javaConnection?.uniqueNameOrNull().orEmpty()
-
-    private fun parseInterfaceName(value: Any?): String = when (value) {
-        is InterfaceName -> value.value
-        is String -> value
-        else -> throw createError(-1, "Properties call failed: invalid interface argument")
-    }
-
-    private fun parsePropertyName(value: Any?): String = when (value) {
-        is PropertyName -> value.value
-        is String -> value
-        else -> throw createError(-1, "Properties call failed: invalid property argument")
-    }
-
-    private fun propertyToVariant(value: Any?): Variant = when (value) {
-        is Variant -> value
-        is Boolean -> Variant(value)
-        is Byte -> Variant(value.toUByte())
-        is UByte -> Variant(value)
-        is Short -> Variant(value)
-        is UShort -> Variant(value)
-        is Int -> Variant(value)
-        is UInt -> Variant(value)
-        is Long -> Variant(value)
-        is ULong -> Variant(value)
-        is Float -> Variant(value.toDouble())
-        is Double -> Variant(value)
-        is String -> Variant(value)
-        is ObjectPath -> Variant(value)
-        is Signature -> Variant(value)
-        is UnixFd -> Variant(value)
-        else -> throw createError(
-            -1,
-            "Properties call failed: unsupported property value type ${value?.let {
-                it::class.simpleName
-            }}"
-        )
-    }
-
-    private fun variantToPropertyValue(variant: Variant): Any = when (variant.peekValueType()) {
-        "b" -> variant.get<Boolean>()
-        "y" -> variant.get<UByte>()
-        "n" -> variant.get<Short>()
-        "q" -> variant.get<UShort>()
-        "i" -> variant.get<Int>()
-        "u" -> variant.get<UInt>()
-        "x" -> variant.get<Long>()
-        "t" -> variant.get<ULong>()
-        "d" -> variant.get<Double>()
-        "s" -> variant.get<String>()
-        "o" -> variant.get<ObjectPath>()
-        "g" -> variant.get<Signature>()
-        "h" -> variant.get<UnixFd>()
-        else -> throw createError(-1, "Properties call failed: unsupported variant value type")
-    }
-
-    private fun ensurePropertiesDispatchRegistered() {
-        if (propertiesDispatchRegistered) return
-        JvmStaticDispatch.register(
-            destination = dispatchDestination,
-            objectPath = objectPath.value,
-            interfaceName = propertiesInterfaceName,
-            methodName = "Get",
-            argCount = 2
-        ) { args ->
-            val iface = parseInterfaceName(args.getOrNull(0))
-            val prop = parsePropertyName(args.getOrNull(1))
-            val property = propertiesByInterface[iface]?.get(prop)
-                ?: throw createError(-1, "Properties.Get failed: unknown $iface.$prop")
-            val reply = PropertyGetReply()
-            JvmCurrentMessageContext.withMessage(reply) {
-                property.getter?.invoke(reply)
-                    ?: throw createError(-1, "Properties.Get failed: no getter for $iface.$prop")
-            }
-            propertyToVariant(reply.payload.firstOrNull())
-        }
-        JvmStaticDispatch.register(
-            destination = dispatchDestination,
-            objectPath = objectPath.value,
-            interfaceName = propertiesInterfaceName,
-            methodName = "Set",
-            argCount = 3
-        ) { args ->
-            val iface = parseInterfaceName(args.getOrNull(0))
-            val prop = parsePropertyName(args.getOrNull(1))
-            val property = propertiesByInterface[iface]?.get(prop)
-                ?: throw createError(-1, "Properties.Set failed: unknown $iface.$prop")
-            val value = args.getOrNull(2)
-            val setValue = if (value is Variant) variantToPropertyValue(value) else value
-            val call = PropertySetCall()
-            val inbound = JvmCurrentMessageContext.current()
-            call.metadata = Message.Metadata(
-                interfaceName = propertiesInterfaceName,
-                memberName = "Set",
-                sender = inbound?.sender?.value,
-                destination = inbound?.destination?.value,
-                path = objectPath.value,
-                valid = true,
-                empty = false
-            )
-            call.payload.add(setValue)
-            JvmCurrentMessageContext.withMessage(call) {
-                property.setter?.invoke(call)
-                    ?: throw createError(-1, "Properties.Set failed: no setter for $iface.$prop")
-            }
-            Unit
-        }
-        JvmStaticDispatch.register(
-            destination = dispatchDestination,
-            objectPath = objectPath.value,
-            interfaceName = propertiesInterfaceName,
-            methodName = "GetAll",
-            argCount = 1
-        ) { args ->
-            val iface = parseInterfaceName(args.getOrNull(0))
-            val props = propertiesByInterface[iface].orEmpty()
-            props.entries.associate { (name, property) ->
-                val reply = PropertyGetReply()
-                JvmCurrentMessageContext.withMessage(reply) {
-                    property.getter?.invoke(reply)
-                        ?: throw createError(
-                            -1,
-                            "Properties.GetAll failed: no getter for $iface.$name"
-                        )
-                }
-                PropertyName(name) to propertyToVariant(reply.payload.firstOrNull())
-            }
-        }
-        propertiesDispatchRegistered = true
-    }
-
-    private fun snapshotInterfaceProperties(interfaceName: String): Map<PropertyName, Variant> {
-        val properties = propertiesByInterface[interfaceName].orEmpty()
-        if (properties.isEmpty()) return emptyMap()
-        return properties.entries.mapNotNull { (name, property) ->
-            val getter = property.getter ?: return@mapNotNull null
-            val reply = PropertyGetReply()
-            val value = JvmCurrentMessageContext.withMessage(reply) {
-                getter(reply)
-                reply.payload.firstOrNull()
-            }
-            PropertyName(name) to propertyToVariant(value)
-        }.toMap()
-    }
-
-    private fun readPropertyValue(property: PropertyVTableItem): Variant {
-        val getter = property.getter ?: error("no getter")
-        val reply = PropertyGetReply()
-        val value = JvmCurrentMessageContext.withMessage(reply) {
-            getter(reply)
-            reply.payload.firstOrNull()
-        }
-        return propertyToVariant(value)
-    }
-
-    // Mirrors native's sd_bus_emit_properties_changed_strv: each named property whose vtable
-    // getter can be read goes into changed_properties with its current value; names without a
-    // readable getter (write-only / value unavailable) fall back to invalidated_properties.
-    private fun resolveChangedProperties(
-        interfaceName: String,
-        propNames: List<PropertyName>
-    ): Pair<Map<PropertyName, Variant>, List<PropertyName>> {
-        val properties = propertiesByInterface[interfaceName].orEmpty()
-        val changed = mutableMapOf<PropertyName, Variant>()
-        val invalidated = mutableListOf<PropertyName>()
-        propNames.forEach { propName ->
-            val property = properties[propName.value]
-            val value = property
-                ?.takeIf { it.getter != null }
-                ?.let { runCatching { readPropertyValue(it) }.getOrNull() }
-            if (value != null) {
-                changed[propName] = value
-            } else {
-                invalidated += propName
-            }
-        }
-        return changed to invalidated
-    }
-
-    override fun emitPropertiesChangedSignal(
-        interfaceName: InterfaceName,
-        propNames: List<PropertyName>
-    ) {
-        val sender = senderName ?: javaConnection?.uniqueNameOrNull()
-        val (changedProperties, invalidatedProperties) =
-            resolveChangedProperties(interfaceName.value, propNames)
-        val payload = listOf(interfaceName, changedProperties, invalidatedProperties)
-        val realConnection = javaConnection
-        if (realConnection == null) {
-            emitLocalOrWire(
-                sender = sender,
-                path = objectPath.value,
-                interfaceName = propertiesInterfaceName,
-                member = propertiesChangedMemberName,
-                signature = propertiesChangedSignature,
-                payload = payload
-            )
-            return
-        }
-        val javaChangedProperties: Map<String, org.freedesktop.dbus.types.Variant<Any?>> =
-            changedProperties.entries.associate { (name, variant) ->
-                val javaPayload = extractVariantPayload(variant)
-                name.value to toJavaVariantValue(javaPayload.signature, javaPayload.value)
-            }
-        val rawSignal = runCatching {
-            realConnection.messageFactory.createSignal(
-                sender ?: realConnection.uniqueNameOrNull(),
-                objectPath.value,
-                propertiesInterfaceName,
-                propertiesChangedMemberName,
-                propertiesChangedSignature,
-                interfaceName.value,
-                javaChangedProperties,
-                invalidatedProperties.map { it.value }
-            )
-        }.getOrElse {
-            throw createError(-1, "emitPropertiesChangedSignal failed: ${it.message}")
-        }
-        runCatching { realConnection.sendMessage(rawSignal) }.getOrElse {
-            throw createError(-1, "emitPropertiesChangedSignal failed: ${it.message}")
-        }
-    }
-
-    override fun emitPropertiesChangedSignal(interfaceName: InterfaceName) {
-        emitPropertiesChangedSignal(interfaceName, emptyList())
-    }
-
-    override fun emitInterfacesAddedSignal() {
-        emitInterfacesAddedSignal(registeredInterfaces.keys.map(::InterfaceName))
-    }
-
-    override fun emitInterfacesAddedSignal(interfaces: List<InterfaceName>) {
-        val sender = senderName ?: javaConnection?.uniqueNameOrNull()
-        val signalPath = LocalObjectManagerRegistry.resolveFor(objectPath.value) ?: objectPath.value
-        val payload = listOf(
-            objectPath,
-            interfaces.associateWith { emptyMap<PropertyName, Variant>() }
-        )
-        val realConnection = javaConnection
-        if (realConnection == null) {
-            emitLocalOrWire(
-                sender = sender,
-                path = signalPath,
-                interfaceName = objectManagerInterfaceName,
-                member = interfacesAddedMemberName,
-                signature = interfacesAddedSignature,
-                payload = payload
-            )
-            return
-        }
-        val javaMap: Map<String, Map<String, org.freedesktop.dbus.types.Variant<Any?>>> =
-            interfaces.associate { it.value to emptyMap() }
-        val rawSignal = runCatching {
-            realConnection.messageFactory.createSignal(
-                sender ?: realConnection.uniqueNameOrNull(),
-                signalPath,
-                objectManagerInterfaceName,
-                interfacesAddedMemberName,
-                interfacesAddedSignature,
-                objectPath.value,
-                javaMap
-            )
-        }.getOrElse {
-            throw createError(-1, "emitInterfacesAddedSignal failed: ${it.message}")
-        }
-        runCatching { realConnection.sendMessage(rawSignal) }.getOrElse {
-            throw createError(-1, "emitInterfacesAddedSignal failed: ${it.message}")
-        }
-    }
-
-    override fun emitInterfacesRemovedSignal() {
-        emitInterfacesRemovedSignal(registeredInterfaces.keys.map(::InterfaceName))
-    }
-
-    override fun emitInterfacesRemovedSignal(interfaces: List<InterfaceName>) {
-        val sender = senderName ?: javaConnection?.uniqueNameOrNull()
-        val signalPath = LocalObjectManagerRegistry.resolveFor(objectPath.value) ?: objectPath.value
-        val payload = listOf(objectPath, interfaces)
-        val realConnection = javaConnection
-        if (realConnection == null) {
-            emitLocalOrWire(
-                sender = sender,
-                path = signalPath,
-                interfaceName = objectManagerInterfaceName,
-                member = interfacesRemovedMemberName,
-                signature = interfacesRemovedSignature,
-                payload = payload
-            )
-            return
-        }
-        val rawSignal = runCatching {
-            realConnection.messageFactory.createSignal(
-                sender ?: realConnection.uniqueNameOrNull(),
-                signalPath,
-                objectManagerInterfaceName,
-                interfacesRemovedMemberName,
-                interfacesRemovedSignature,
-                objectPath.value,
-                interfaces.map { it.value }
-            )
-        }.getOrElse {
-            throw createError(-1, "emitInterfacesRemovedSignal failed: ${it.message}")
-        }
-        runCatching { realConnection.sendMessage(rawSignal) }.getOrElse {
-            throw createError(-1, "emitInterfacesRemovedSignal failed: ${it.message}")
-        }
-    }
-
-    override fun addObjectManager(): Resource {
-        val local = LocalObjectManagerRegistry.register(objectPath.value, dispatchDestination)
-        registrations += local
-        // Wire backend: also advertise ObjectManager for over-the-wire introspection/serving.
-        val wireResource = if (wireSignalEmitter != null) {
-            WireServeRegistry.registerObjectManager(dispatchDestination, objectPath.value)
-                .also { registrations += it }
-        } else {
-            null
-        }
-        return ActionResource {
-            local.release()
-            wireResource?.release()
-            registrations.remove(local)
-            wireResource?.let { registrations.remove(it) }
-        }
-    }
-
-    override fun currentlyProcessedMessage(): Message =
-        JvmCurrentMessageContext.current() ?: signalFromMetadata(Message.Metadata())
-
-    override fun addVTable(interfaceName: InterfaceName, vtable: List<VTableItem>): Resource {
-        registeredInterfaces[interfaceName.value] =
-            (registeredInterfaces[interfaceName.value] ?: 0) + 1
-        val interfaceProperties = propertiesByInterface.getOrPut(interfaceName.value) {
-            mutableMapOf()
-        }
-        val addedPropertyNames = mutableListOf<String>()
-        vtable.forEach { item ->
-            val property = item as? PropertyVTableItem ?: return@forEach
-            interfaceProperties[property.name.value] = property
-            addedPropertyNames += property.name.value
-        }
-        if (addedPropertyNames.isNotEmpty()) {
-            ensurePropertiesDispatchRegistered()
-        }
-        val resources = mutableListOf<Resource>()
-        resources += LocalManagedObjectsRegistry.registerInterface(
-            destination = dispatchDestination,
-            objectPath = objectPath.value,
-            interfaceName = interfaceName.value
-        ) {
-            snapshotInterfaceProperties(interfaceName.value)
-        }
-        // Wire backend (epic #93 phase 4): capture the vtable metadata so the owned connection can
-        // route, introspect and serve this object to EXTERNAL callers over the bus, not just
-        // in-process. The dbus-java/stub paths (wireSignalEmitter == null) don't need this.
-        if (wireSignalEmitter != null) {
-            resources += WireServeRegistry.registerVTable(
-                destination = dispatchDestination,
-                path = objectPath.value,
-                interfaceName = interfaceName.value,
-                vtable = vtable
-            )
-        }
-        resources += vtable.mapNotNull { item ->
-            val method = item as? MethodVTableItem ?: return@mapNotNull null
-            val callback = method.callbackHandler ?: return@mapNotNull null
-            val argCount = countTopLevelTypes(method.inputSignature?.value)
-            JvmStaticDispatch.register(
-                destination = dispatchDestination,
-                objectPath = objectPath.value,
-                interfaceName = interfaceName.value,
-                methodName = method.name.value,
-                argCount = argCount
-            ) { args ->
-                val call = MethodCall().also {
-                    val inbound = JvmCurrentMessageContext.current()
-                    it.metadata = Message.Metadata(
-                        interfaceName = interfaceName.value,
-                        memberName = method.name.value,
-                        sender = inbound?.sender?.value,
-                        destination = inbound?.destination?.value,
-                        path = objectPath.value,
-                        valid = true,
-                        empty = args.isEmpty()
-                    )
-                    it.payload.addAll(args)
-                }
-                JvmCurrentMessageContext.withMessage(call) {
-                    callback(call)
-                }
-                if (method.hasNoReply) return@register Unit
-                val reply = call.sentReply ?: call.awaitReply(asyncReplyTimeoutMs)
-                    ?: throw createError(
-                        -1,
-                        "No reply sent for ${interfaceName.value}.${method.name.value} within ${asyncReplyTimeoutMs}ms"
-                    )
-                JvmStaticDispatch.DispatchResult(reply)
-            }
-            ActionResource {
-                JvmStaticDispatch.unregister(
-                    destination = dispatchDestination,
-                    objectPath = objectPath.value,
-                    interfaceName = interfaceName.value,
-                    methodName = method.name.value,
-                    argCount = argCount
-                )
-            }
-        }
-        registrations.addAll(resources)
-        return ActionResource {
-            resources.forEach { it.release() }
-            registrations.removeAll(resources)
-            val current = registeredInterfaces[interfaceName.value] ?: 0
-            if (current <= 1) {
-                registeredInterfaces.remove(interfaceName.value)
-            } else {
-                registeredInterfaces[interfaceName.value] = current - 1
-            }
-            if (addedPropertyNames.isNotEmpty()) {
-                val props = propertiesByInterface[interfaceName.value]
-                addedPropertyNames.forEach { props?.remove(it) }
-                if (props.isNullOrEmpty()) {
-                    propertiesByInterface.remove(interfaceName.value)
-                }
-            }
-        }
-    }
-
-    override fun createSignal(interfaceName: InterfaceName, signalName: SignalName): Signal =
-        signalFromMetadata(
-            Message.Metadata(
-                interfaceName = interfaceName.value,
-                memberName = signalName.value,
-                sender = senderName ?: javaConnection?.uniqueNameOrNull(),
-                path = objectPath.value,
-                valid = true,
-                empty = true
-            )
-        ) { emitSignal(it) }
-
-    override fun emitSignal(message: Signal) {
-        val interfaceName = message.interfaceName?.value
-            ?: throw createError(-1, "emitSignal failed: missing interface name")
-        val signalName = message.memberName?.value
-            ?: throw createError(-1, "emitSignal failed: missing signal name")
-        val path = message.path?.value ?: objectPath.value
-        // Prefer the declared signature (correct even for empty collections, which carry no
-        // runtime element to infer from); fall back to value inference only when no usable
-        // declared signature is available. Mirrors the method-call body path.
-        val signature = bodySignature(
-            message.payload.toList(),
-            message.declaredBodySignature.toString()
-        ) { "emitSignal failed: unsupported signal payload type" }
-        val args = message.payload.map(::toJavaSignalValue).toTypedArray()
-        val sender = message.sender?.value ?: senderName ?: javaConnection?.uniqueNameOrNull()
-        val realConnection = javaConnection
-        if (realConnection == null) {
-            emitLocalOrWire(
-                sender = sender,
-                path = path,
-                interfaceName = interfaceName,
-                member = signalName,
-                signature = signature,
-                payload = message.payload
-            )
-            return
-        }
-        val rawSignal = runCatching {
-            realConnection.messageFactory.createSignal(
-                sender ?: realConnection.uniqueNameOrNull(),
-                path,
-                interfaceName,
-                signalName,
-                signature.ifEmpty { null },
-                *args
-            )
-        }.getOrElse {
-            throw createError(-1, "emitSignal failed: ${it.message}")
-        }
-        runCatching { realConnection.sendMessage(rawSignal) }.getOrElse {
-            throw createError(-1, "emitSignal failed: ${it.message}")
-        }
-    }
-
-    override fun release(): Unit = run {
-        registrations.forEach { it.release() }
-        registrations.clear()
-    }
-}
